Add README docs and cleanup C codegen
