+++ tool/src/c/structs.rs
@@ -1,0 +1,92 @@
+use std::fmt::Write;
+use std::{collections::HashMap, fmt};
+
+use diplomat_core::ast;
+use indenter::indented;
+
+use super::types::gen_type;
+
+pub fn gen_struct<W: fmt::Write>(
+    custom_type: &ast::CustomType,
+    in_path: &ast::Path,
+    env: &HashMap<ast::Path, HashMap<String, ast::ModSymbol>>,
+    out: &mut W,
+) -> fmt::Result {
+    match custom_type {
+        ast::CustomType::Opaque(opaque) => {
+            writeln!(out, "typedef struct {} {};", opaque.name, opaque.name)?;
+        }
+
+        ast::CustomType::Struct(strct) => {
+            write!(out, "typedef struct {} {{", strct.name)?;
+            let mut class_body_out = indented(out).with_str("    ");
+            for (name, typ, _) in strct.fields.iter() {
+                writeln!(&mut class_body_out)?;
+                gen_field(name, typ, in_path, env, &mut class_body_out)?;
+            }
+            writeln!(out)?;
+            writeln!(out, "}} {};", strct.name)?;
+        }
+
+        ast::CustomType::Enum(_) => {}
+    }
+
+    Ok(())
+}
+
+pub fn gen_field<W: fmt::Write>(
+    name: &str,
+    typ: &ast::TypeName,
+    in_path: &ast::Path,
+    env: &HashMap<ast::Path, HashMap<String, ast::ModSymbol>>,
+    out: &mut W,
+) -> fmt::Result {
+    gen_type(typ, in_path, env, out)?;
+    write!(out, " {};", name)?;
+
+    Ok(())
+}
+
+pub fn gen_method<W: fmt::Write>(
+    method: &ast::Method,
+    in_path: &ast::Path,
+    env: &HashMap<ast::Path, HashMap<String, ast::ModSymbol>>,
+    out: &mut W,
+) -> fmt::Result {
+    match &method.return_type {
+        Some(ret_type) => {
+            gen_type(ret_type, in_path, env, out)?;
+        }
+
+        None => {
+            write!(out, "void")?;
+        }
+    }
+
+    write!(out, " {}(", method.full_path_name)?;
+    let mut params_to_gen = method.params.clone();
+    if let Some(param) = &method.self_param {
+        params_to_gen.insert(0, param.clone());
+    }
+
+    for (i, param) in params_to_gen.iter().enumerate() {
+        if i != 0 {
+            write!(out, ", ")?;
+        }
+
+        if param.ty == ast::TypeName::StrReference {
+            write!(
+                out,
+                "const char* {}_data, size_t {}_len",
+                param.name, param.name
+            )?;
+        } else {
+            gen_type(&param.ty, in_path, env, out)?;
+            write!(out, " {}", param.name)?;
+        }
+    }
+
+    writeln!(out, ");")?;
+
+    Ok(())
+}
